simulation: remove all activity dependencies for indexation of 'delivery' category
They were useless since 'delivery' is indexed in ZODB, and this also fixes
a bug causing local build to fail in the following conditions:
- root document moved to a state that triggers a builder,
  whereas there's no simulation tree yet
- the builder select method has a condition on the root simulation movement

An example is building of task reports, when the ERP is overloaded.

The reason was that in some cases, ERP5 tried to set 2 tags on the same
reindexing activity (built: in updateMovementCollection & expand: in
_updateSimulation), but there's actually no support for multi-valued tags
and for CMFActivity, default activate parameters (here expand:) have lower
precedence (see ActivateObject.activate). So another possible fix is to add
built: to _localBuild after_tag.

This commit also renames expand: into build: